feat : modal 분리, icon component 수정
