+++ package.json
@@ -50,6 +50,7 @@
     "@types/jest": "27.5.1",
     "@types/react": "^17.0.43",
     "@types/react-native": "^0.67.3",
+    "@types/react-native-vector-icons": "^6.4.10",
     "@types/react-test-renderer": "^17.0.1",
     "@types/styled-components": "^5.1.24",
     "@types/styled-components-react-native": "^5.1.3",
